Version 0.34_qa --------------------------------------     + added Compression Algorithm     + "fix compression" now uses the new compression ability     + WOS is generating Texture Names again     + Text Lists 0x88 do now get renamed properly

+++ trunk/fullsimpe/SimPe Interfaces/IPackedFile.cs
@@ -48,7 +48,7 @@ namespace SimPe.Interfaces.Files
 		/// <summary>
 		/// Returns the Size of the File
 		/// </summary>
-		int Signature
+		ushort Signature
 		{
 			get;
 		}
